Added exception throws for Booking service

+++ LifeStyle_Club_Facility_Booking/src/main/java/com/lsc/mvc/service/BookingServiceImpl.java
@@ -12,6 +12,10 @@ import org.assertj.core.util.Arrays;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.lsc.mvc.exception.BookingNotFound;
+import com.lsc.mvc.exception.FacilityNotFound;
+import com.lsc.mvc.exception.ResourceDefinitionInvalid;
+import com.lsc.mvc.exception.UserNotFound;
 import com.lsc.mvc.model.Booking;
 import com.lsc.mvc.model.Facility;
 import com.lsc.mvc.repository.BookingRepository;
@@ -21,9 +25,15 @@ public class BookingServiceImpl implements BookingService {
 
 	@Resource
 	private BookingRepository bRepo;
+	@Resource
+	private FacilityService fService;
+	@Resource
+	private UserService uService;
 	
 	@Override
-	public Booking setNewBookingNum(Booking b) {
+	public Booking setNewBookingNum(Booking b) throws BookingNotFound {
+		if (b == null) throw new BookingNotFound("Booking object provided cannot be null");
+		
 		// Getting Current Max userId
 		Integer newId = bRepo.getBookingIdMax() + 1;
 		
@@ -36,31 +46,38 @@ public class BookingServiceImpl implements BookingService {
 	
 	@Override
 	@Transactional
-	public Booking addBooking(Booking b) {
+	public Booking addBooking(Booking b) throws BookingNotFound {
+		if (b == null) throw new BookingNotFound("Booking object provided cannot be null");
 		return bRepo.saveAndFlush(b);
 	}
 	
 	@Override
 	@Transactional
-	public Booking getBooking(String bNum) {
-		return bRepo.getBookingByBookingNumber(bNum);
+	public Booking getBooking(String bNum) throws BookingNotFound {
+		Booking b = bRepo.getBookingByBookingNumber(bNum);
+		if (b == null) throw new BookingNotFound("Booking number provided is invalid");
+		else return b;
 	}
 
 	@Override
 	@Transactional
-	public Booking updateBooking(Booking b) {
+	public Booking updateBooking(Booking b) throws BookingNotFound {
+		if (b == null) throw new BookingNotFound("Booking object provided cannot be null");
 		return bRepo.saveAndFlush(b);
 	}
 
 	@Override
 	@Transactional
-	public Void removeBooking(String bNum) {
+	public Void removeBooking(String bNum) throws BookingNotFound {
+		if (getBooking(bNum)==null) throw new BookingNotFound("Booking number provided is invalid");
 		bRepo.delete(getBooking(bNum));
 		return null;
 	}
 	
 	@Override
-	public ArrayList<Booking> getBookingListSelected(ArrayList<Facility> facilityListSelected){
+	public ArrayList<Booking> getBookingListSelected(ArrayList<Facility> facilityListSelected) throws ResourceDefinitionInvalid {
+		if (facilityListSelected.size() == 0) throw new ResourceDefinitionInvalid("Facility List provided is empty");
+		
 		// Create New ArrayList of Bookings
 		ArrayList<Booking> bList = new ArrayList<Booking>();
 		
@@ -68,10 +85,14 @@ public class BookingServiceImpl implements BookingService {
 		for (Facility f:facilityListSelected) {
 			bList.addAll(bRepo.getBookingListByFacilityNumber(f.getFacilityNumber()));
 		}
+	
 		return bList;
 	}
 	@Override
-	public ArrayList<Booking> getBookingListSingle(ArrayList<Booking> bookingListSelected, String fNum) {
+	public ArrayList<Booking> getBookingListSingle(ArrayList<Booking> bookingListSelected, String fNum) throws ResourceDefinitionInvalid, FacilityNotFound {
+		if (bookingListSelected.size() == 0) throw new ResourceDefinitionInvalid("Booking List provided is empty");
+		if (fService.getFacility(fNum)==null) throw new FacilityNotFound("Facility number provided is invalid");
+		
 		// Create New ArrayList of Bookings
 		ArrayList<Booking> bList = new ArrayList<Booking>();
 		
@@ -79,11 +100,14 @@ public class BookingServiceImpl implements BookingService {
 		for (Booking b: bookingListSelected) {
 			if (b.getFacilityNumber().equals(fNum)) bList.add(b);
 		}
+		
 		return bList;
 	}
 	
 	@Override
-	public List<String> getAvailableSlots(LocalDate slotDate, String fNum) {
+	public List<String> getAvailableSlots(LocalDate slotDate, String fNum) throws FacilityNotFound {
+		if (fService.getFacility(fNum)==null) throw new FacilityNotFound("Facility number provided is invalid");
+		
 		// Retrieve ArrayList of Bookings
 		ArrayList<Booking> bList = bRepo.getBookingListByFacilityNumberAndDate(fNum, slotDate);
 		
@@ -92,28 +116,36 @@ public class BookingServiceImpl implements BookingService {
 		aList.add("0900"); aList.add("1000"); aList.add("1100"); aList.add("1200"); aList.add("1300"); aList.add("1400");
 		aList.add("1500"); aList.add("1600"); aList.add("1700"); aList.add("1800"); aList.add("1900"); aList.add("2000");
 		
-		// Convert Booking Blocks into Hourly Slots
-		ArrayList<String> bStrList = new ArrayList();
-		for(String str:aList)
-			for(Booking b:bList) 
-				if (str.compareTo(b.getSlotTimeStart()) >= 0 && str.compareTo(b.getSlotTimeEnd()) < 0) bStrList.add(str);
-		
-		// Add Any Possible Slot Not in Booked Slots into Output List 
+		// Create ArrayList for Output
 		List<String> outList = new ArrayList();
-		for(String str:aList)
-			if (!bStrList.contains(str)) outList.add(str);
+		
+		if (bList.size() > 0) {
+			// Convert Booking Blocks into Hourly Slots
+			ArrayList<String> bStrList = new ArrayList();
+			for(String str:aList)
+				for(Booking b:bList) 
+					if (str.compareTo(b.getSlotTimeStart()) >= 0 && str.compareTo(b.getSlotTimeEnd()) < 0) bStrList.add(str);
+			
+			// Add Any Possible Slot Not in Booked Slots into Output List 
+			for(String str:aList)
+				if (!bStrList.contains(str)) outList.add(str);
+		} else {
+			outList = aList;
+		}
 		
 		// Return Output List
 		return outList;
 	}
 	
 	@Override
-	public ArrayList<Booking> getBookingListByUserNumAndDate(String uNum, LocalDate dateStart, LocalDate dateEnd) {
+	public ArrayList<Booking> getBookingListByUserNumAndDate(String uNum, LocalDate dateStart, LocalDate dateEnd) throws UserNotFound {
+		if (uService.getUser(uNum)==null) throw new UserNotFound("User number provided is invalid");
 		return bRepo.getBookingListByUserNumAndDate(uNum, dateStart, dateEnd);
 	}
 	
 	@Override
-	public ArrayList<Booking> getBookingListByUserNum(String uNum) {
+	public ArrayList<Booking> getBookingListByUserNum(String uNum) throws UserNotFound {
+		if (uService.getUser(uNum)==null) throw new UserNotFound("User number provided is invalid");
 		return bRepo.getBookingListByUserNum(uNum);
 	}
 }
